requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
cmake==3.25.2
docopt==0.6.2
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
pipreqs==0.4.11
requests==2.28.2
smmap==5.0.0
urllib3==1.26.14
yarg==0.1.9
